feat: fix export html build
