+++ apps/web/app/[lng]/(home)/page.next.tsx
@@ -3,11 +3,6 @@ import { Metadata, ResolvingMetadata } from 'next/types'
 
 import { HomePage } from '@web/components/HomePage/HomePage'
 import { getServerTranslation } from '@web/i18n/server'
-import { languages } from '@web/i18n/settings'
-
-export async function generateStaticParams() {
-  return languages.map((lng) => ({ lng }))
-}
 
 export async function generateMetadata(
   props: ScriptProps,
