+++ apps/web/app/[lng]/(home)/page.spec.tsx
@@ -2,7 +2,7 @@ import { screen } from '@testing-library/react'
 
 import { renderApp } from '@sdlgr/test-utils'
 
-import Page, { generateMetadata, generateStaticParams } from './page.next'
+import Page, { generateMetadata } from './page.next'
 
 jest.mock('next/headers', () => ({
   cookies: () => ({
@@ -16,20 +16,6 @@ describe('[lng] route - Page', () => {
     const items = await screen.findAllByRole('presentation')
     expect(items).toHaveLength(2)
     expect(baseElement).toMatchSnapshot()
-  })
-})
-
-describe('[lng] route - static params', () => {
-  it('should generate static params successfully', async () => {
-    const params = await generateStaticParams()
-    expect(params).toEqual([
-      {
-        lng: 'en',
-      },
-      {
-        lng: 'es',
-      },
-    ])
   })
 })
